build/base: install megacli from local packages in Debian plateforms
Background: hwraid.le-vert.net is *very* often down and block CI
process.
The package is currently in base role so as we do for Red Hat / CentOS
roles, let's use a .deb package from hwraid.le-vert.net stored on
github/enovance/edeploy-roles/files as we already do for Red Hat /
CentOS with a zip file.

This patch aims to:
* delete add_megacli_repository function (not used anymore)
* update install_megacli function to install megacli from a local debian
  package
  The patch in edeploy-roles is:
https://github.com/enovance/edeploy-roles/commit/c76ad319795fa727144993275103557de4af19ec
* move the unzip binary check in install_megacli for Red Hat / CentOS
  since add_megacli_repository is deleted.