edeploy.prof: Using pci=bfsort by default
We have a case where two hosts of the same kind reports the
same hardware devices in a different order.

As as result, eth0,1,2,3 from host1 is mapped on eth0,2,3,1 on the 2nd
one.

I've been investigating the difference an nothing except the naming is
different. This is not an issue for eDeploy but it is for health/cardiff
when trying to group system by similar properties.

This patch is adding pci=bfsort to insure than pci devices are always
reported in the same order.